requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
Appium-Python-Client==2.11.1
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
PyJWT==2.7.0
PySocks==1.7.1
pytest==7.2.1
pytest-xdist==3.2.1
python-dotenv==0.21.1
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==4.0.0
wsproto==1.2.0